Added some content over
The main content has been added consisting of two rows, with the first row containing some plain text and the second row of the main divided into 3 equal columns. The placeholder text has been added as well. Now to begin styling with CSS.
Also, aligned the copyright text in the footer over to the right.