Finished the homepage structure
I added the horizontal lines to enclose the navigation bar, so that it can help differentiate from the rest of the contents in the webpage, as well as adding margin to space out the navigation links. I also centered most of the main content and added some placeholder images. (I don't have any pictures added yet, so I added alternate text for now.) Finally, both the logo at the header and footer are adjusted so that it can fit and resize properly when displaying the webpage via mobile OS emulation (like responsive tools).