Tidy the webpage
Instead of using div, I use article HTML tags for the main content. Also I added some border lines for both the promotional and column rows so that I can use them as a reference when displaying the contents. This is temporary, so I will remove them prior to final submission.