Published some changes
First, I added the hover transition effect, such as changing the background color for the navigation links on hover and click, changing the background color for the promotional banner from gold to dark goldenrod on hover, and causing the columns to become more transparent on hover. 
Next, I adjusted the images for the column elements so that they can resize properly when vieweing the homepage in responsive mobile mode.
Finally, instead of "Find a location" for the third column, I chose "Business Hours" because I want to keep the columns evenly shaped while viewing the webpage in mobile responsive mode.